Split up cli.js
Move the actual implementation into lib/cli.js to avoid top-level
returns. xo@^0.17.0 cannot parse the file with those returns present.

Throw errors from lib/cli.js and write them to stderr in cli.js before
exiting. XO doesn't appreciate process.exit() calls in modules.

Note that the error for invalid Babel config is now written to stderr,
which seems like an improvement to me.
