+++ test/cli.js
@@ -70,7 +70,7 @@ function execCli(args, opts, cb) {
 }
 
 test('disallow invalid babel config shortcuts', function (t) {
-	execCli('es2015.js', {dirname: 'fixture/invalid-babel-config'}, function (err, stdout) {
+	execCli('es2015.js', {dirname: 'fixture/invalid-babel-config'}, function (err, stdout, stderr) {
 		t.ok(err);
 
 		var expectedOutput = '\n  ';
@@ -78,7 +78,7 @@ test('disallow invalid babel config shortcuts', function (t) {
 		expectedOutput += ' See ' + chalk.underline('https://github.com/avajs/ava#es2015-support') + ' for allowed values.';
 		expectedOutput += '\n';
 
-		t.is(stdout, expectedOutput);
+		t.is(stderr, expectedOutput);
 		t.end();
 	});
 });
